requirements.txt
defusedxml==0.4.1
Django==1.7
django-filter==0.9.1
django-guardian==1.2.5
djangorestframework==3.0.3
jsonfield==1.0.0
Markdown==2.5.2
PyYAML==3.11
six==1.9.0
